chore(file structure):refactored file structure with testing in mind
