+++ src/components/pages/FAQ/FAQ.tsx
